Implementation of Field Injection in activities, Usage of Modules and Provides
The default way for a class to request its dependencies is via its @Inject-annotated constructor.

That ensures that the dependencies are present when the object is fully constructed, as is normal Java best practice. There are, however, a few other options.

The @Inject annotation is also valid on fields and methods. Dagger will set @Inject fields after the constructor is invoked. Similarly, it will call @Inject methods, supplying the parameters as if they were dependencies specified in the constructor.

Modules are a way of telling Dagger how to provide dependencies from the dependency graph. These are typically high level dependencies that you aren't already contributing to the dependency graph through the @Inject constructor annotation